rename v2check to v2fix
