+++ tools/v2fix/_stage/ddtracetypes/ddtracetypes.go.golden
