+++ tools/v2fix/_stage/traceidstring/traceidstring.go.golden
