+++ tools/v2fix/_stage/tracerstructs/tracerstructs.go.golden
