+++ tools/v2fix/v2fix/known_change.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2023 Datadog, Inc.
 
-package v2check
+package v2fix
 
 import (
 	"context"
