+++ tools/v2fix/v2fix/v2fix_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2023 Datadog, Inc.
 
-package v2check_test
+package v2fix_test
 
 import (
 	"context"
@@ -14,7 +14,7 @@ import (
 	"path"
 	"testing"
 
-	"github.com/DataDog/dd-trace-go/tools/v2check/v2check"
+	"github.com/DataDog/dd-trace-go/tools/v2fix/v2fix"
 
 	"golang.org/x/tools/go/analysis"
 	"golang.org/x/tools/go/analysis/analysistest"
@@ -38,10 +38,10 @@ func (c V1Usage) Fixes() []analysis.SuggestedFix {
 	return nil
 }
 
-func (c V1Usage) Probes() []v2check.Probe {
-	return []v2check.Probe{
-		v2check.IsFuncCall,
-		v2check.HasPackagePrefix("gopkg.in/DataDog/dd-trace-go.v1"),
+func (c V1Usage) Probes() []v2fix.Probe {
+	return []v2fix.Probe{
+		v2fix.IsFuncCall,
+		v2fix.HasPackagePrefix("gopkg.in/DataDog/dd-trace-go.v1"),
 	}
 }
 
@@ -54,35 +54,35 @@ func (c V1Usage) String() string {
 }
 
 func TestV1Usage(t *testing.T) {
-	c := v2check.NewChecker(&V1Usage{
+	c := v2fix.NewChecker(&V1Usage{
 		ctx: context.Background(),
 	})
 	c.Run(testRunner(t, "v1usage"))
 }
 
 func TestV1ImportURL(t *testing.T) {
-	c := v2check.NewChecker(&v2check.V1ImportURL{})
+	c := v2fix.NewChecker(&v2fix.V1ImportURL{})
 	c.Run(testRunner(t, "v1importurl"))
 }
 
 func TestDDTraceTypes(t *testing.T) {
 	t.Skip()
-	c := v2check.NewChecker(&v2check.DDTraceTypes{})
+	c := v2fix.NewChecker(&v2fix.DDTraceTypes{})
 	c.Run(testRunner(t, "ddtracetypes"))
 }
 
 func TestWithServiceName(t *testing.T) {
-	c := v2check.NewChecker(&v2check.WithServiceName{})
+	c := v2fix.NewChecker(&v2fix.WithServiceName{})
 	c.Run(testRunner(t, "withservicename"))
 }
 
 func TestTraceIDString(t *testing.T) {
-	c := v2check.NewChecker(&v2check.TraceIDString{})
+	c := v2fix.NewChecker(&v2fix.TraceIDString{})
 	c.Run(testRunner(t, "traceidstring"))
 }
 
 func TestTracerStructs(t *testing.T) {
-	c := v2check.NewChecker(&v2check.TracerStructs{})
+	c := v2fix.NewChecker(&v2fix.TracerStructs{})
 	c.Run(testRunner(t, "tracerstructs"))
 }
